Address Steven's comments and improve comments.  Refs #84.
The most significant change is the removal of cases in the Store
code that would reference either RHS or &RHS multiple times.  As
Steven observed, they could produce incorrect results if the target
of the store "influenced" the value of RHS (e.g. if the target
appears as part of an array index expression in RHS).